Refactoring previously migrated settings commands
